inject rpath in short ld commands
The short symlinks can be utilized without the activation scripts. A use case
where this might be helpful is if one was cross compiling and thus the activation
scripts would be primarily concerned with the HOST or TARGET architecture.  Standards
for FLAGS that apply to the BUILD are a bit varied  FLAGS_FOR_BUILD is used in some
cases but there are others that completely ignore those.

The symlinked short-name tools require CONDA_PREFIX/include and the CONDA_PREFIX/lib rpath
at a minimum.  For tools that need those flags, inject them in small wrapper scripts.
They are installed with the PREFIX and thus will be modified at install time to have the
installed PREFIX hardcoded.

In the case of binutils, we have the additional advantage when using wrapper scripts
that we can tweak LDFLAGS to the idiosyncrasies of each ld implementation.  Of note,
LDFLAGS are typically passed into the gcc driver.  However, there are numerous build
scripts that call ld directly with LDFLAGS and having the '-Wl' (send to the linker)
option for the gcc driver causes annoying errors.  Also, ld.gold doesn't understand
-rpath-link, it expects --rpath-link.